Minor fixes
Fixed some data type issues with the library itself.
Worked on a bit of documentation.

I'm currently thinking about how I might reimplement strToInt() in a different manner so as to support types such as 'long long' and their unsigned counterparts.

+++ MyBasics.h
@@ -22,7 +22,7 @@ int fDiscardLine(FILE *stream);
 
 size_t getStr(char **str, const char delim, const size_t length, FILE *stream);
 
-int getStrStdin(char **str, const size_t length);
+size_t getStrStdin(char **str, const size_t length);
 
 int indexOf(const char *str, const char letter, const size_t offset);
